requirements.txt
Django
Pillow
kavenegar
djangorestframework
twilio
psycopg2
django-cors-headers
